fix for verify feature
